Gen_AI/10_mcp_server/requirements.txt
mcp
google-generativeai
python-dotenv
